refactor(Core.JSONParser): nest all JSONParser types under Core.JSONParser; rename files; extract Apple URL constants
### New Core.JSONParser sub-namespace

Anchor file Core.JSONParser.swift declares the sub-namespace. The legacy
top-level JSONCrawler namespace marker is gone; all types live under
Core.JSONParser.

Types relocated:
- AppleJSONCrawlerEngine          -> Core.JSONParser.Engine
  + JSONCrawlerError              -> Core.JSONParser.Engine.Error
- JSONContentFetcher              -> Core.JSONParser.ContentFetcher
  + JSONFetcherError              -> Core.JSONParser.ContentFetcher.Error
- AppleJSONToMarkdown             -> Core.JSONParser.AppleJSONToMarkdown
- MarkdownToStructuredPage        -> Core.JSONParser.MarkdownToStructuredPage
- RefResolver                     -> Core.JSONParser.RefResolver
- AppleJSONAPITitleFetcher        -> Core.JSONParser.AppleJSONAPITitleFetcher
- WKWebViewTitleFetcher           -> Core.JSONParser.WKWebViewTitleFetcher
- CompositeTitleFetcher           -> Core.JSONParser.CompositeTitleFetcher

Errors qualify Swift.Error in conformance lists to break self-reference.

### File renames per <Namespace>+<Type>.swift convention

- AppleJSONCrawlerEngine.swift -> Core.JSONParser+Engine.swift
- JSONContentFetcher.swift     -> Core.JSONParser+ContentFetcher.swift
- AppleJSONToMarkdown.swift    -> Core.JSONParser+AppleJSONToMarkdown.swift
- MarkdownToStructuredPage.swift -> Core.JSONParser+MarkdownToStructuredPage.swift
- RefResolver.swift            -> Core.JSONParser+RefResolver.swift
- RefResolverFetchers.swift    -> Core.JSONParser+RefResolverFetchers.swift
- JSONCrawler.swift            -> Core.JSONParser.swift  (anchor)

### External callers swept

- Core/Core+Crawler.swift          (production)
- CLI/Commands/Command+ResolveRefs.swift (production)
- Search/Search+IndexBuilder.swift (production)
- CoreTests/AppleJSONToMarkdownTests.swift
- CoreTests/CrawlerTests.swift
- CoreTests/RefResolverTests.swift

All AppleJSONToMarkdown, RefResolver, MarkdownToStructuredPage references
qualified through Core.JSONParser.<type>.

### Hardcoded Apple URL extraction (user request, screenshot showed 164 hits)

Production code now uses Shared.Constants.BaseURL.* exclusively. Two new
constants added:

- appleArchiveDocs    = '\(Shared.Constants.BaseURL.appleArchive)documentation/'
- appleTutorialsData  = 'https://developer.apple.com/tutorials/data'
- appleTutorialsDocs  = appleTutorialsData + '/documentation'

Production sites swept (no behaviour change):

- AppleJSONToMarkdown.swift    : 3 sites -> appleDeveloper / appleDeveloperDocs / appleTutorialsData
- Search/Search+IndexBuilder.swift : 2 sites -> appleDeveloperDocs
- TUI/Models/ArchiveEntry.swift    : 1 site  -> appleArchiveDocs
- Core/Core+ArchiveGuideCatalog.swift : baseURL derived from appleArchiveDocs
- Core/Core+TechnologiesIndexFetcher.swift : 2 sites -> appleTutorialsDocs / appleDeveloper
- Core/SampleCodeDownloader.swift  : 1 site  -> appleDeveloper
- Availability/Availability+Fetcher.swift : default apiBaseURL -> appleTutorialsDocs
- Availability target gains SharedConstants dep

Tests and docs/README literals untouched (they document the actual URLs
on purpose).

### Verification

- xcrun swift build (clean): 0 warnings, 0 errors
- xcrun swift test (clean):  1300/1300 pass
